added Tags Posts

+++ app/Models/Post.php
@@ -2,10 +2,12 @@
 
 namespace App\Models;
 
+use Carbon\Carbon;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Cviebrock\EloquentSluggable\Sluggable;
+use Illuminate\Support\Str;
 
 class Post extends Model
 {
@@ -14,15 +16,15 @@ class Post extends Model
     const IS_DRAFT = 0;
     const IS_PUBLIC = 1;
 
-    protected $fillable = ['title', 'content'];
+    protected $fillable = ['title', 'content', 'date'];
 
     public function category()
     {
-        return $this->hasOne(Category::class);
+        return $this->belongsTo(Category::class);
     }
     public function author()
     {
-        return $this->hasOne(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
     public function tags()
     {
@@ -58,17 +60,24 @@ class Post extends Model
     }
     public function remove()
     {
-        Storage::delete('uploads/' . $this->image);
+        $this->removeImage();
         $this->delete();
+    }
+
+    public function removeImage()
+    {
+        if ($this->image != null) {
+            Storage::delete('uploads/' . $this->image);
+        }
     }
     public function uploadImage($image)
     {
         if ($image == null) {
             return;
         }
-        Storage::delete('uploads/' . $this->image);
-        $filename = str_random(10) . '.' . $image->extension();
-        $image->saveAs('uploads', $filename);
+        $this->removeImage();
+        $filename = Str::random(10) . '.' . $image->extension();
+        $image->storeAs('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
@@ -131,7 +140,38 @@ class Post extends Model
             return $this->setFeatured();
         }
     }
+    public function setDateAttribute($value)
+    {
+        $date = Carbon::createFromFormat('d/m/y', $value)->format('Y-m-d');
+        // $this->attributes['date'] = strtolower($value);
+        $this->attributes['date'] = $date;
+    }
+
+    public function getCategoryTitle()
+    {
+        // if ($this->category != null) {
+        //     return $this->category->title;
+        // }
+        // return 'Нет категории!';
+        return ($this->category != null)
+            ?   $this->category->title
+            :   'Нет категории!';
+    }
+
+    public function getTagsTitles()
+    {
+        // if (!$this->tags->isEmpty()) {
+        //     return implode(', ', $this->tags->pluck('title')->all());
+        // }
+        // return 'Нет тегов!';
+        return (!$this->tags->isEmpty())
+            ?   implode(', ', $this->tags->pluck('title')->all())
+            : 'Нет тегов!';
+    }
 }
+
+
+
 // $post = Post::find(1);
 // $post->category->title;
 // $post->tags;
